pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-review-env"
version = "1.0.0"
description = "OpenEnv environment for AI pull request code review evaluation"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "openai>=1.0",
]

[project.scripts]
server = "code_review_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["code_review_env", "code_review_env.server"]
package-dir = { "code_review_env" = ".", "code_review_env.server" = "server" }
